Graph data structure
The resources page is now populated with all the visually appealing, user friendly features which allows user to assign rooms and servers which user can add manually to the teams for assigned period of time.
Also the user can add the tasks to be completed as well
This features makes the application makes allocating sprints possible.